If you refresh home and look at the network panel you will see cors errors
This is because Home is sending include credentials in the fetch request but the demo is not adding that in the response headers (so it is saying it doesn't accept them). The data still shows up in the list but a bunch of errors are raised. This should resolve that.