Update shortcut and github installer so that main = latest release
Versioned branches should always point to a versioned manifest with a shortcut that represents the version. The installer link should point to that versioned manifest. Main changes over time and if someone wants to run latest (without pinning via DOS) the main branch is updated over time to point to the latest version of these samples through main branch manifests.