Added support for inline-view (where the manifest is defined in apps.json), added a new fdc3 view, and version updates
Added support for inline-view (where the manifest is defined in apps.json), An example of this is the fdc3 workbench entry in the apps.json file. It doesn't have a separate view json file as it is inline. Launch and the interop broker were updated to support inline views.

Updated the native example that now includes an embedded view that also includes the fdc3 workbench.

Updated dos files and readmes.