Added two buttons for showing home and toggling notification
Added two buttons for showing home and toggling notification and added example of interop broker hooks for validating connection to broker or actions performed once connected to the broker (e.g. you could use this hook to deny a view from within a platform joining or publishing to a group).