Update the npm commands to support mac and windows (#80)
the npm run client commands now work on a mac as well as windows. Hitting Ctrl+C or Command+C on the terminal will also close the openfin application.

the npm run kill command now works on a mac as well as windows.

Readme updated to reflect that mac can be used for development.