Updates to support workspace management apis
Added support for the SaveMenu instead of SavePage. Updated the menu.ts file to take all the default menu options and use logic to insert before/after, delete or update label to help with upgrades. Dropped description from the workspace saving in home and added logic to remove logic if it is the current workspace. Work still to do on updating a result if a workspace is switched/deleted/renamed from the browser menu and the result is already listed in home.