Workspace/vnext (7.x) into main (#100)
* Initial next commit

* Update bundle

* Update to latest and rebuild

* Update minimum runtime

* Updates to support workspace management apis

Added support for the SaveMenu instead of SavePage. Updated the menu.ts file to take all the default menu options and use logic to insert before/after, delete or update label to help with upgrades. Dropped description from the workspace saving in home and added logic to remove logic if it is the current workspace. Work still to do on updating a result if a workspace is switched/deleted/renamed from the browser menu and the result is already listed in home.

* Update to resolve type issue with resize-observer

* Npm install latest and rebuild. Update Native Window Integration Project

* Updating version number.

* Update to version 7.3.10

* Merge in license.md

* Update v6 to v7 references and show where to override workspaces

* Consistent version in shortcuts