[DEVOP-71] extend github workflow to publish to multiple locations
Until now, `workspace-starter` project has been publishing their HOWTOs
to GitHub Pages. A need has arisen to publish to other locations as well
to broaden the availability of the HOWTOs for customers that may not be
able to access GitHub Pages for various reasons.

This commit refactors the current GitHub Actions Workflow to allow
publishing to multiple locations. It also refactors the packaging script
to allow custom location and location path setup when preparing URLs in
various HOWTO manifests.