Add additional manifestType examples
Add additional examples to show:

* How you can launch a headless window to do background processing
* How you can launch a classic window from home (inline definition or via a manifest json file)
* How you can launch the desktop browser if needed (maybe a site that is only compatible with an older desktop browser).

You need to run npm run setup to update and build the common folder (which has the hidden window code).

The only sample that demonstrates this is the customize-workspace sample as it is giving an example of launching custom manifest types. New manifest types:

* window
* inline-window
* desktop-browser

The manifests in customize-workspace have been updated to opt into displaying these manifest types.