Dev/john/endpoint use consistency (#163)
* Initial commit

Rather than using endpoints for sharing and share overrides, and custom url settings for app endpoints and then a different setting related to where workspaces, pages and page bounds are stored (through a platform-storage mechanism).

The appsSourceUrl is still supported and if detected it will use the old behaviour. endpointIds can point to a defined endpoint (they also support urls directly at which point it will try and use fetch. We recommend using an endpoint definition as it gives you more options).

We now allow you to specify cache in seconds and minutes.

You can now define an endpoint that uses a javascript module. The second manifest includes an example of using a module from the common folder for storing the workspaces, pages and page bounds. The main manifest still uses the default storage mechanism for pages and workspaces.

Did some tidy up and refactoring but still need to tidy up docs and a few more tweaks.

* Update where platform override comes from and better naming.

* Updates after feedback