Dev/john/settings (#209)
* Updated to support fetching settings from a rest endpoint

You can override a manifest setting with a rest endpoint and if auth is setup this settings call could be personalised to the user that has logged in.

* Valid manifest host defaults to host of platform but additional options can be specified via a root json file.

Valid manifest host defaults to host of platform but additional options can be specified via a root json file. I didn't want the domains hardcoded in the js. This gives someone the freedom to simply override/update the json file.

* linting

* updating based on feedback

* Additional tweak

Co-authored-by: Martyn Janes <martyn@obany.com>