Added ability for connected apps to specify an app to launch if they are not running (#215)
Connected apps now have the option of specify an app entry in the snapshot if they are not running. This is something that would be launched by the platform. The saving app can decide what to pass in order to help rehydrate itself.

Example manifest entry:

"connectionProvider": {
			"connectionId": "workspace-connection",
			"supportedActions": [
				"show-home",
				"show-store",
				"show-dock",
				"hide-home",
				"hide-store",
				"minimize-dock"
			],
			"connections": [
				{
					"identity": { "uuid": "openfin-dotnet6-wpf-test-harness" },
					"validatePayload": false,
					"connectionTypes": [{ "type": "appSource" }, { "type": "snapshotSource" }, { "type": "actions" }]
				},
				{
					"identity": { "uuid": "openfin-dotnet4-winform-test-harness" },
					"validatePayload": false,
					"connectionTypes": [{ "type": "appSource" }, { "type": "snapshotSource" }, { "type": "actions" }]
				}
			]
		},

The supported manifest types would need to include the followng:

"connection",
"inline-external"