Keeping payload validation off
The payload verification endpoint returns true so it can be enabled by turning it on if you want to see the flow but it will means more hops with no benefit. So now in place to be turned on for testing in preparation for when you build a real validation endpoint.