Added an app endpoint mapping pattern
Added the ability to specify that you wish to map apps from one format to the platform's format. Added an endpoint module that looks at the shape of fdc3 1.2 and 2.0 apps and takes the required information and maps it onto our platform app shape.

Also added ability to support inline-external for cases where the fdc3 2.0 type native was specified.