Improved conditions
Decoupled conditions from importing sharing and auth.
Added function to conditions to allow the platform to register internal conditions.
Added function to themes to find out if the platform supports light and dark.
Added logging to conditions
Add condition "themed" and used it as a condition of the toggle theme button.