Adding IntelliSense support to manifests and settings.json
To improve the dev experience we are generating a json schemma from the settings shapes.ts file. We also have a wrapper schema for use in manifest that imports the generated schema. Settings in a production environment would be server generated but for learning we wanted to make it easier to update settings without always having to go the documentation. A small change was done to the browser shapes to support this initiative and it means we will be adding richer comments to these configurable option.