Add IntelliSense to settings.json and manifest customSettings (#289)
* Adding IntelliSense support to manifests and settings.json

To improve the dev experience we are generating a json schemma from the settings shapes.ts file. We also have a wrapper schema for use in manifest that imports the generated schema. Settings in a production environment would be server generated but for learning we wanted to make it easier to update settings without always having to go the documentation. A small change was done to the browser shapes to support this initiative and it means we will be adding richer comments to these configurable option.

* Updated code documentation to improve schema intellisense

Also did some updates that came out of looking the existing manifest/setting.json file and how they used the options.

* Update linting rules to ignore generated schema file

* Applying feedback

* Added some mandatory settings.

* Updated action shapes to use relative app shapes and cleaned up types

* Resolving review based points

* Updated based on pull request review