Updates based on pull request review
Some tidying up. Plus removed the name extension to Platform App. Right now name was only being used for specific fdc3 use cases (open and the fdc3 mappers). The name property could be confusing if we are not fully utilizing it so for now the support for name has been placed where it is needed.