Added opinionated fdc3.open support (#296)
* Added opinionated fdc3.open support

Also fixed a bug where you could specify an appId when raising an intent and if the app wasn't listed but other apps supported that intent then the intent picker (or a single app but not the intended one) would be launched.

Added documentation related to fdc3.open support.

* Updates based on pull request review

Some tidying up. Plus removed the name extension to Platform App. Right now name was only being used for specific fdc3 use cases (open and the fdc3 mappers). The name property could be confusing if we are not fully utilizing it so for now the support for name has been placed where it is needed.