Update the example auth to support entitled auth
The example auth module now includes an optional setting to pass customData to the specified login window. The custom data includes a dummy list of users and roles and a key with which to store the user.

The auth module now includes a supporting endpoint which can be used in the endpoint module definition to specify the key to look up the current user and then a role map which can be keyed off the role from the current user to exclude certain things (menu options) from the settings and certain things from the returned apps. In a real app this would all be done on the server side so only the data the user is entitled to would be returned. The demo module uses client side logic to give a portable and easy to use example that doesn't depend on server capabilities. I also created a new developer actions module that lets you plug in the option to open the dev tools for a view, window or platform from the browser. The second.manifest.fin.json example now lets you pick a developer or sales person role and updates the menu and app options accordingly. Documentation has also been updated.