Update to support specifying a scheme preference even with a single palette
If you have a single palette relying on the title of the theme to include light is quite restrictive. We are utilising the default to specify a preference for the palette/palettes. If someone is specifying a single palette that is light based then the default should be light to let the platform know the icons to load if they decide to use our placeholders. Updating our platform theme to support default across both modes and updated settings and manifests to demonstrate new settings.