Improved startup process
Added more logging.

Now we only return an interop client after the bootstrap process to ensure the broker has been setup and we only trigger things in platform after the platform has indicated it is ready (rather than just after the await).

In the lifecycle we have more logging and we make a copy of the array before notifying subscribers (as there is an edge case where a subscriber removing an entry from the list after running will stop execution of other subscribers.

Added example of how to get the interop client using the lifecycle functions in the analytics implementation in the dev module. Added some extra logic in that example dev module to only publish interop events that come from the browser (so we are not looping events generated from the module).

Enabled this module in the main manifest.