Moved notifications to promise vs await
This will ensure it doesn't wait for the notification centre when it can continue bootstrapping.

Added locale options (the platform doesn't react to it right now because all strings are passed via settings but it is useful to provide to modules should the platform wish to provide that information).

The helper functions could be reduced so that there was one getThemeInfo (or something) function vs the 3/4 theme related functions we currently have.

Need to test via branch before merging to main.