Added wider intent target support (#353)
* Added wider intent target support

Added support for window and inline-window manifest type to the broker as a supported intent target. Added example of why you might want to use this and the private app setting (so apps do not show in home and store but can be intent targets). New contact dashboard example added to common by adding an intent snapshot launcher window which can take custom settings to support the launching and modifying of a snapshot. This shows targeting specific context groups and dynamic replacement of tokens in a snapshot (provided an example of something that might not support intents but can have the context passed to a querystring (it could have been custom data).