Dev/john/fdc3 (#392)
* More updates to broker related code

* Updates to add a second intent picker option

This second version shows instances. This branch is still a work in progress.

* Updating picker

* Saving progress

* Refactor with instance support

Still a work in progress

* linting

* More improvements

* More updates

* Added instance information

* Updates to allow focusing on the target of an intent.

* Update images and launching of intents when it is a single intent

* Update way of tracking intent registrations to make it better with async

Dropped map to an object and tidied up the code. You can now launch from a saved workspace and it tracks launched and then closed instances just like it does if you launched them directly.

* Updated fdc3 call app to 2.0 and only bring existing instances to front.

* Applying feedback from pull request review

* Add support for customConfig mapping in fdc3 mappers and updated changelog

* Re-enable analytics

* Use custom config to assign a unique id if the metadata indicates it

* Update helper to allow fdc3 2.0 and allow the passing of viewOptions

* Upgrade to latest v12

Also added a ts 5 ignore config entry as we are still pinned to v4 but updated tooling can trigger a warning error for a change coming to ts 5.5

* Make launch the only way of launch apps and support timeout for intents

Plus wrap rvm info call as that will not work on a mac.

* Update bug in launch logic and fix readme

We were generating a new name for the identity but not assigning it to the manifest which would mean the container would generate it's own identity for the view or window. Also cleaned up some logging from the view logic (likely used for testing).

* Added option for supporting registered intent handlers that are not linked to an app

* Added support for bringing unregistered app instances to front and more error handling

* Cut down the number of instances of intent apps that are listed to make it easier to get to grips with.