Updated to support wider default options and omit menu entries
We now support:

defaultWindowOptions
defaultViewOptions
defaultPageOptions

in a browserProvider configuration. We have backwards compatibility for the old windowOptions (we kept the main manifest in the old format and updated the others to the new for this commit to show it working in both cases). defaultWindowOptions now opens up more configuration options.

We also have a new menu options setting that lets you specify whether or not you want to include the default menu entries provided by the workspace platform.

Updated changelog and documentation.