Decoupled home from app logic, extended home provider settings
The support for showing apps in home now follows the pattern used for adding workspace and page support to home. It is a pluggable module that you can use, remove or replace with your own. It has the same features as before but you now need to specify it in the integrationsProvider settings.

Updated settings accepted by the homeProvider configuration. It now accepts a wider range of settings to match the settings that you could specify via code.

Updated documentation and tightened types. Updated manifests and settings.json to include the new apps integration.