Decoupled home from app logic, extended home provider settings (#421)
* Decoupled home from app logic, extended home provider settings

The support for showing apps in home now follows the pattern used for adding workspace and page support to home. It is a pluggable module that you can use, remove or replace with your own. It has the same features as before but you now need to specify it in the integrationsProvider settings.

Updated settings accepted by the homeProvider configuration. It now accepts a wider range of settings to match the settings that you could specify via code.

Updated documentation and tightened types. Updated manifests and settings.json to include the new apps integration.

* Updated app module to not include workspace package, integration updates and more logging

- Updated apps to only import the type for workspace npm modules this makes the integration module much smaller and it also impacts home and triggering a search when enter is hit when the same module is loaded twice into memory.
- Updated workspace and pages integration module to match the provideId pattern where we use the definition id so that they always match.
- Updated home with a log of the number of results returned.
- Updated platform override so that there is a lot more logging related to page and workspace actions and there was a bug where a page/workspace change wouldn't trigger a lifecycle event if custom storage was being used.

* Update pack.json to reflect new modules that have been built

A lot of opinionated pieces related to workspaces, apps and pages in home is now broken out into optional modules. This pack file includes those additional modules in case they are useful. The fdc3 bundle is no longer included because fdc3 support is now built in.

* Improve logging for handled results

* Fix linting

---------

Co-authored-by: Martyn Janes <martyn@obany.com>