Added support for richer menu definitions and added new page composite for menus
We now extend the menu item template to expose the extra settings (so you can now specify icon and submenu).

There is an optional new composite module (enabled in the default manifest) that provides a dynamic menu and a collection of three actions. This lets you show an existing page in the current window. Unless:

- The window doesn't allow pages to be added (at which point it will launch the page in a new window)
- The page is already open on the current window or another window. The logic will make the page active and bring the window to the foreground.

There is also a delete page capability.