Added logic to either use the single intent or the selected intent (when there are many)
The logic was not populating the intent object that would be empty if raiseIntentByContext was selected.

If there was only one intent available after checking and the passed intent did not specify a name (because it was by context) then we should update the intent name before continuing (as it will be used).

If there were multiple intents available and the passed intent did not specify a name (e.g. because it was by context) then we should update the intent with the intent selected by the user.