Updated example module and secondmanifest (#459)
* Updated example module and secondmanifest

The structure of the manifest was updated to use urls instead of endpoints and the developer menu options was moved from inline menu options to a menu module. The example auth module needed to be updated to reflect this change. We extended it to support menu modules and to also support urls regardless of their format (platform app array, fdc3 1.2 or fdc3 2.0). We extended the fdc3 1.2 definition to support categories that was introduced in fdc3 2.0. These get mapped to tags in a platform app. We still support specifying tags in an fdc3 1.2 directory if provided but the schema was aligned with FDC3 2,0.

* Improve types

---------

Co-authored-by: Martyn Janes <martyn@obany.com>