Workspace/v13.0.0 (#487)
Workspace 13 samples provide a number of changes. The samples have been simplified and we no longer include server (expressjs) code where it isn't needed (we use an npm server to show that they are just there to serve the files). The entire repo is now typescript strict true enabled and we have tightened our linting rules. 

We have also renamed customize workspace to workspace-platform-starter to make it clearer that it is an example of a workspace platform (kitchen sink but still an example) rather than just an example of customizing the UI of a workspace.