Added initial fav support (#538)
We have had discussions around this in the past. Here is something that would allow endpoints to be specified to store/retrieve/remove an object that represents a favorite. There is a set of understood types of thing you can favorite and this can be passed to modules so that actions, integrations and possibly other sections could utilize it.

---------

Co-authored-by: Martyn Janes <martyn@obany.com>