Added logic for awaiting instanceId on broker connection
When opening an app via fdc3 we should await the connection to the broker to get the endpointId which acts as a unique instance id that cannot be specified by an app.

FDC3 2.0 open and get info tests now pass (although there seems to be something with the last open test that can occasionally fail so needs investigation to see if it is the test or the broker).