Notifications support
First attempt at putting a pattern around how a platform can use notifications across multiple teams within a platform.

The idea is to introduce a level of isolation between modules and allowing modules to have be grouped and have settings to specify whether they should be a part of a platform or just show up in the all tab. Plus the options around whether the icon from a module should be enforced when it comes to notifications.

This pattern also means that modules do not need to bundle notification code from the workspace module as it is injected. It also ensures all modules use the same version.

Todo:

- Docs
- Changelog
- The addEventListener needs to support all the overloaded functions
- The removeEventListener logic needs tweaking as it isn't working and needs to support the overloaded logic.
- Experimental code was added to the apps and pages integration for testing purposes. This needs to be removed.