This needs one more tweak.
There seems to be an issue if you have a single module and you reference it twice with different module entry settings. It can be that you just have to create a new module for isolation but will be reviewing in the morning to see if you can re-use one module with two bootstrap lifecycle events.

Not ready for merging just yet.