Notifications support (#564)
Initial example of putting a pattern around how a platform can use notifications across multiple teams within a platform.

The idea is to introduce a level of isolation between modules and allowing modules to be grouped and have settings to specify whether they should be a part of a platform or just show up in the all tab. Plus the options around whether the icon from a module should be enforced when it comes to notifications.

This pattern also means that modules do not need to bundle notification code from the workspace module as it is injected. It also ensures all modules use the same version.

- Notification Client created and Notification Provider supports a getNotificationClient function that can be passed options.
- Docs updated
- Changelog updated
- Example lifecycle module added demonstrating notification client use.
- Main manifest updated with example lifecycle module reference but disabled.