Added additional error handling around window positioning strategy and option to turn it off
Added additional error handling around trying to determine if a window is visible and what it's bounds are because during an automation related test where many windows are being launched and closed there is a chance that a reference to a window no longer exists when you try to get whether it is showing or it's position.

As window layout strategy is really for an end user it serves little benefit during testing where the goal is to run automation tests as quickly as possible. Therefore the option to disable the window positioning strategy has been added.