Add new init options support to composite page module (#578)
The page composite module now supports launching/showing a page by specifying init params (similar to querystring params). This will either launch the page or bring it to front if it was already loaded.

It is launched using the following structure:
fins://domain/yourmanifest.fin.json?$$action=show-page&$$payload=eyAicGFnZUlkIjogImIwY2UxNTg3LTM2ZDAtNGRlZC05ZGU3LTlmNmQyYjc1OGYyNyIgfQ==

A localhost example:

fin://localhost:8080/manifest.fin.json?$$action=show-page&$$payload=eyAicGFnZUlkIjogImIwY2UxNTg3LTM2ZDAtNGRlZC05ZGU3LTlmNmQyYjc1OGYyNyIgfQ==

The payload is a base64 encoded string of a json object containing a pageId e.g.

{ "pageId": "b0ce1587-36d0-4ded-9de7-9f6d2b758f27" }

This pageId is unlikely to exist on your setup but it is just an example.

The module is enabled via the initProvider configuration in manifest.fin.json but the configuration can be disabled or removed if you don't want this support.

	"initOptionsProvider": {
		"modules": [
			{
				"enabled": true,
				"id": "show-page",
				"title": "Init Options Show Page",
				"url": "http://localhost:8080/js/modules/composite/pages.bundle.js",
				"data": {
					"supportedActions": ["show-page"]
				}
			}
		]
	},
The launchPage helper method passed to modules has been updated so that it now accepts a pageId instead of requiring modules to look up the page object and pass it.

the condition function (used to check to see if a certain condition is true) has also been moved from integration helpers to the base module helpers object to allow more modules to check for conditions.

Changelog has been updated to include the changes.