Initial look at using popup window
The popup window works I just need to update the positional logic as it is slightly off center.

There is some investigation needed because awaiting the popup window means calls from the popup window to find fdc3 instances does not resolve because the broker is still waiting for a response from the popup window.