Updates to style and update intent picker logic
The border width/height should be 100% not 99% (validated with Martyn) so this is being pushed out across samples.

The interop broker can use the popup window api but there is a known issue (being resolved in a future runtime) where a popup window running on localhost cannot use the interop api. A todo was added to swap localhost with 127.0.0.1 just for the intent picker url until that runtime release is out.

Updated documentation and updated positioning logic.