Added support for endpoint support for manifest fetching and tidied helpers
Removed the theme related functions that were at the root of module helpers as they are now available through getThemeClient. The TS Types are there to indicate they no longer exist and the changelog has been updated to reflect the change. This now follows our getXClient pattern.

New support for specifying endpoints for fetching the manifest json url specified in an app definition. A platform can specify an endpoint called manifest-get if it wants to override the default fetch. If a platform wants a specific behavior for a specific app then they can set an endpoint called manifest-get-app-id.

The default fetch implementation for endpoints has been enhanced. Now you can have an endpoint defined with different options for fetch but no url and have the request contain the url when calling request/response etc. This is used to allow platform devs to provide additional fetch options specific to their platform but have the manifest url passed without the need to have a custom endpoint module created.

e.g.

{
	"id": "manifest-get",
	"type": "fetch",
	"options": {
	      "method": "GET"
		}
	}