Applied suggested typo changes in readme and applied fixes. (#604)
* Applied suggested typo changes in readme and applied fixes.

* Tidy up share

add applyWorkspace logic:

If you have not loaded anything or have opened and then closed windows then there should be no prompt on whether or not you wish to load a workspace. We do not override the skipPrompt if it has been passed to applyWorkspace but we determine whether or not it should be skipped on first load.