Added a check around the getAppAssetInfo call
This call checks to see if an app asset is already available and if it is the expected version. If it isn't then it tries to download the app asset. If the app asset doesn't exist an error is thrown so this adds logic to handle and log that information.