Empower platform devs to create an unrestricted notification service (#682)
This option allows a platform dev to nominate one or more lifecycle modules (or other types) to have an unrestricted notification client (no icon, or in platform tab enforcement and the notifications are not scoped to prefix id or module id). This is achieved by listing the module in the notification clients section of the notification provider and setting it's restricted setting to false (default is true).