Salesforce updates
Added option to easily decide if browser pages launched from home should be assigned to a context group and whether it should be the same context group or rotated.

Added additional intent target support to the salesforce example.

Added an example of how to add a salesforce app as an intent target to workspace platform starter.