Added support for configuring the interop broker to not send context to the sender
Something that has come up is use cases where you have a context listener that never wants any context that has been sent from the same webpage and instance.

There are ways of doing this so you ignore context messages sent from yourself (and we updated the doc for this) but now we also have a configuration option so you can specify that you do not want the originator of a context message to receive it.