Added platform override example and fixes
Added an example platform override module that can update saved urls if the app definition changes or show an access denied page if the user no longer has access to the app.

Updated the launch snapshot logic so that view ids are consistently returned.

Updated page logic to ensure that page view ids with guids are regenerated to prevent clashes (where a view is moved from a page to another window and then the page is relaunched).

Added docs and updated the changelog.