Bump dos version and fix menu delete capability
DOS needed to be bumped to v19.1 and the menu.ts file had a bug where specifying a menu option deletion would delete the menu entry and entries that followed as the splice was not restricted to the single entry.

Updated changelog for WPS and how to customize browser menu documentation with more examples.