Added protection against log modules logging against themselves when using createLogger
A log module is a log sink. It is passed a createLogger function as it is a module but if used it would result in the logged messages being passed back to itself (as well as other log sinks).

An additional check has been done when initializing log modules to prevent the log sink linked logger being passed. A local logger is returned that logs just to the console (for debugging purposes) when called from a log module.