Fixed two issues and added two examples
Fixed an issue where the popup window "announce" example on the custom toolbar example was not passing the customData to the popup window.

Fixed the OpenPage logic as the logic was checking for a custom button instead of a global context menu button so it would not open a page when clicked.

Added an example of adding a menu option to debug the platform.

Added an example of renaming a page dynamically and then saving the changes.