Added additional example of server auth
Added a second manifest with an example where the provider is not visible by default and preload scripts are used to determine if the right url is loaded in the hidden window.

Also added an additional scenario where the server auth may use redirects and a redirect may become stuck and the provider.html is never loaded.