Updated snap logic
Snap will not try and launch multiple instances of the same appId/instanceId via launchExternal or app asset if it exists. This may be child windows that are being tracked by snap so it shouldn't result in multiple launches of the same exe (the native app should look at supporting snapshot source).