Syncing v20.1 with the latest changes in main. (#765)
* Added context listener to update UI when context is received from BBG Launchpad (#761)

* Added context listener to update UI when context is received from BBG Launchpad.

* Added context listener to listen for messages from Launchpad

* Fixed linting and added a Java App example (#762)

* Updated snap example to allow easier configuration of snap options. (#763)

Also updated other parts of code due to restrictions of later versions of typescript.

* Updated snap logic (#764)

Snap will not try and launch multiple instances of the same appId/instanceId via launchExternal or app asset if it exists. This may be child windows that are being tracked by snap so it shouldn't result in multiple launches of the same exe (the native app should look at supporting snapshot source).

---------

Co-authored-by: Ameet Jayawant <ameetj85@gmail.com>